feat(ir): Change get_block_idx return type to UINT64 and add flatten tests
Change get_block_idx operator to return UINT64 instead of INT32 for
better support of large block indices. Add Python API wrapper and test
cases for flatten_call_expr pass with get_block_idx in control flow.

Changes:
- Change DeduceBlockGetBlockIdxType to return UINT64 scalar type
- Add get_block_idx Python API wrapper in block_ops.py
- Add test for get_block_idx in if condition expressions
- Add test for get_block_idx in for range expressions

+++ src/ir/op/block_ops/memory.cpp
@@ -58,8 +58,8 @@ TypePtr DeduceBlockGetBlockIdxType(const std::vector<ExprPtr>& args,
                                    const std::string& op_name) {
   CHECK(args.size() == 0) << "The operator " << op_name << " requires no arguments, but got " << args.size();
 
-  // get_block_idx returns INT32 scalar
-  return std::make_shared<ScalarType>(DataType::INT32);
+  // get_block_idx returns UINT64 scalar
+  return std::make_shared<ScalarType>(DataType::UINT64);
 }
 
 TypePtr DeduceBlockLoadType(const std::vector<ExprPtr>& args,
